% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteMerge.R
\name{add_class_info}
\alias{add_class_info}
\title{add_class_info}
\usage{
add_class_info(
  annot_table,
  annot_table_file =
    "~/Project/00_IBD_project/Data/20240524_B001_B033_data_processing/merge_annotation_manual_24626.xlsx",
  previous_annot_file =
    "~/Project/00_IBD_project/Data/20240402_CD_non_IBD_analysis/annot_table_with_class_update_240403.xlsx"
)
}
\arguments{
\item{annot_table_file}{annotation table}

\item{previous_annot_file}{previous annotation table}
}
\value{
a data frame with classfire information
}
\description{
add_class_info
}
\examples{
test3 <- add_class_info(annot_table_file = '~/Project/00_IBD_project/Data/20240524_B001_B033_data_processing/merge_annotation_manual_24626.xlsx',
                       previous_annot_file = '~/Project/00_IBD_project/Data/20240402_CD_non_IBD_analysis/annot_table_with_class_update_240403.xlsx')
}
