% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteMerge.R
\name{compare_previous_annotation}
\alias{compare_previous_annotation}
\title{compare_previous_annotation}
\usage{
compare_previous_annotation(
  annot_table,
  annot_table_file = ".",
  previous_annot_file = "."
)
}
\arguments{
\item{annot_table_file}{new annotation file}

\item{previous_annot_file}{previous annotation file}
}
\value{
a data frame with previous check result. The column "previous_check" is the result of previous annotation check. "-1" means no previous annotation, "0" means not match, "1" means match.
}
\description{
compare previous annotation with current annotation, and add a column with previous check result
}
\examples{
test <- compare_previous_annotation(annot_table_file = '~/Project/00_IBD_project/Data/20240402_CD_non_IBD_analysis/annot_table_with_class_update_240403.xlsx',
                                    current_annot_file = '~/Project/00_IBD_project/Data/20240524_B001_B033_data_processing/merge_annotation_manual_24626.xlsx')
}
\author{
Zhiwei Zhou
}
