% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteMerge.R
\name{correct_annot_table}
\alias{correct_annot_table}
\title{correct_annot_table}
\usage{
correct_annot_table(
  annot_table_file =
    "~/Project/00_IBD_project/Data/20240524_B001_B033_data_processing/merge_annotation_manual_24626.xlsx",
  correct_table_file =
    "~/Project/04_package/00_Database/DoddLib/01_compound_RT_lib/05_version_231220/corrected_compound_table_231220.xlsx"
)
}
\arguments{
\item{annot_table_file}{annotation table file}

\item{correct_table_file}{corrected compound table file}
}
\value{
a data frame with corrected metabolite information
}
\description{
correct the metabolite infor in the annotation table using the corrected compound table
}
\examples{
test2 <- correct_annot_table(annot_table_file = '~/Project/00_IBD_project/Data/20240524_B001_B033_data_processing/merge_annotation_manual_24626.xlsx',
                            correct_table_file = '~/Project/04_package/00_Database/DoddLib/01_compound_RT_lib/05_version_231220/corrected_compound_table_231220.xlsx')
}
\author{
Zhiwei Zhou
}
