% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteMerge.R
\name{generate_annot_table_manual}
\alias{generate_annot_table_manual}
\title{generate_annot_table_manual}
\usage{
generate_annot_table_manual(
  annot_table_file,
  previous_annot_file = ".",
  correct_table_file = ".",
  dir_path = "."
)
}
\arguments{
\item{annot_table_file}{annotation table file}

\item{previous_annot_file}{previous annotation file}

\item{correct_table_file}{corrected compound table file}

\item{dir_path}{'.'}
}
\description{
generate_annot_table_manual
}
\examples{
generate_annot_table_manual(annot_table_file = '~/Project/00_IBD_project/Data/20240524_B001_B033_data_processing/merge_annotation_manual_24626.xlsx',
                           previous_annot_file = '~/Project/00_IBD_project/Data/20240402_CD_non_IBD_analysis/annot_table_with_class_update_240403.xlsx',
                           correct_table_file = '~/Project/04_package/00_Database/DoddLib/01_compound_RT_lib/05_version_231220/corrected_compound_table_231220.xlsx')
}
\author{
Zhiwei Zhou
}
