% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunOtherWorkflow.R
\name{generate_input_from_msp}
\alias{generate_input_from_msp}
\title{generate_input_from_msp}
\usage{
generate_input_from_msp(
  msp_file = "~/Project/05_coworker/Kazuma/241014/ms2_spec.msp",
  dir_path = "."
)
}
\arguments{
\item{msp_file}{path to msp file}

\item{dir_path}{path to directory where input files will be saved}
}
\value{
input files for annotate_metabolite
}
\description{
generate input files from msp file for annotate_metabolite
}
\examples{
generate_input_from_msp(msp_file = '~/Project/05_coworker/Kazuma/241014/ms2_spec.msp',
                       dir_path = '~/Project/05_coworker/Kazuma/241014/')
}
