% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteId.R
\name{initialize_annotation_parameter_class}
\alias{initialize_annotation_parameter_class}
\title{initialize_annotation_parameter_class}
\usage{
initialize_annotation_parameter_class(
  object = NULL,
  ms1_file = "data.csv",
  ms2_type = c("mzML", "mzXML", "mgf", "msp"),
  path = ".",
  lib = c("dodd", "msdial", "peptide", "gnps_bile_acid", "gnps_acyl_amides",
    "gnps_acyl_esters", "all_public"),
  ce = c("20", "10", "40"),
  column = c("hilic", "c18"),
  polarity = c("positive", "negative"),
  is_rt_score = TRUE,
  is_ms2_score = TRUE
)
}
\arguments{
\item{column}{'hilic', 'c18'}
}
\value{
a object of RawMsParameterClass, with slots:
 \describe{
 \item{name}-{parameter set name}
 \item{para_peak_detection}-{parameter of peak detection}
 }
}
\description{
generate a AnnotationParameterClass for processing raw mass spec data
}
\examples{
\dontrun{
object <- initialize_annotation_parameter_class(path = '~/Project/00_IBD_project/Data/20230327_raw_data_processing_test/DemoData/',
                                                lib = 'dodd',
                                                ce = '20',
                                                column = 'hilic',
                                                polarity = 'positive',
                                                is_rt_score = TRUE,
                                                is_ms2_score = TRUE)
}
}
\author{
Zhiwei Zhou
}
