% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteMerge.R
\name{merge_one_modes}
\alias{merge_one_modes}
\title{merge_one_modes}
\usage{
merge_one_modes(
  path = ".",
  column = c("hilic", "c18"),
  polarity = c("positive", "negative")
)
}
\arguments{
\item{path}{'.'}

\item{column}{'hilic', 'c18'}

\item{polarity}{'positive', 'negative'}
}
\description{
merge multiple annotation using different database and parameter. Note: The input table should be same
}
\author{
Zhiwei Zhou
}
