% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMetaboliteMerge.R
\name{plot_merge_ms2}
\alias{plot_merge_ms2}
\title{plot_merge_ms2}
\usage{
plot_merge_ms2(path = ".", annot_merge_file = "merge_annotation_pos.xlsx")
}
\arguments{
\item{path}{'.' the dirname of 4 folders that contains metabolite identification result}

\item{annot_merge_file}{"merge_annotation_pos.xlsx"}
}
\description{
plot ms2 of merged table for manual check
}
\author{
Zhiwei Zhou
}
